requirements.txt
alabaster==0.7.10
amqp==1.4.9
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
anyjson==0.3.3
asn1crypto==0.24.0
astroid==1.6.1
astropy==2.0.3
atomicwrites==1.2.1
attrs==17.4.0
Babel==2.5.3
beautifulsoup4==4.6.0
billiard==3.3.0.23
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
blinker==1.4
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
celery==3.1.26.post2
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.5.11
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.1.4
cssselect==1.0.3
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
Django==2.0.6
django-celery==3.2.2
django-debug-toolbar==1.9.1
django-environ==0.4.5
django-mssql==1.8
django-pandas==0.6.0
django-pyodbc-azure==2.0.4.1
django-slack==5.12.1
django-truncate==0.1
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
feedfinder2==0.0.4
feedparser==5.2.1
filelock==2.0.13
flake8==3.5.0
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
gevent==1.2.2
glob2==0.6
greenlet==0.4.12
h5py==2.7.1
heapdict==1.0.0
holidays==0.9.4
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==5.0.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
jieba3k==0.35.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.4
jupyterlab-launcher==0.10.2
kombu==3.0.37
lazy-object-proxy==1.3.1
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
mysqlclient==1.3.12
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
newsapi-python==0.2.2
newspaper3k==0.2.6
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.14.0
numpydoc==0.7.0
oauthlib==2.0.7
odo==0.5.1
olefile==0.45.1
openpyxl==2.4.10
packaging==16.8
pandas==0.22.0
pandas-market-calendars==0.15
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prometheus-client==0.4.0
prompt-toolkit==1.0.15
psutil==5.4.3
py==1.5.2
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.4
pylint==1.8.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyowm==2.8.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
python-highcharts==0.4.1
pytz==2017.3
PyWavelets==0.5.2
pywin32==222
pywinpty==0.5
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.13.0
requests-file==1.4.3
requests-oauthlib==0.8.0
robobrowser==0.5.3
rope==0.10.7
ruamel-yaml==0.15.46
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
sqlparse==0.2.4
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.8.2
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
tldextract==2.2.0
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
tweepy==3.6.0
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
web.py==0.40.dev1
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.3.0
zict==0.1.3
